refactor: Service 계층 통일 - 문제 요구사항에 따라 FileXXXService, BasicXXXService를 생성하였습니다. - 다만, Service 계층의 로직은 Basic, File, JCF 계열 모두 동일하다고 판단되어 코드 복사/붙여넣기 방식으로 구현하였습니다. - 실제로는 JCF와 File 계열을 사용하지 않을 예정이므로, 해당 클래스들에는 @Deprecated 어노테이션을 추가하였습니다.
