+++ src/main/java/com/sprint/mission/discodeit/config/AppConfig.java
@@ -10,28 +10,24 @@ import com.sprint.mission.discodeit.repository.file.FileUserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
 import com.sprint.mission.discodeit.service.MessageService;
 import com.sprint.mission.discodeit.service.UserService;
-import com.sprint.mission.discodeit.service.jcf.JCFChannelService;
-import com.sprint.mission.discodeit.service.jcf.JCFMessageService;
-import com.sprint.mission.discodeit.service.jcf.JCFUserService;
+import com.sprint.mission.discodeit.service.basic.BasicChannelService;
+import com.sprint.mission.discodeit.service.basic.BasicMessageService;
+import com.sprint.mission.discodeit.service.basic.BasicUserService;
 
 public class AppConfig {
-
     public ChannelRepository channelRepository = new FileChannelRepository();
     public UserRepository userRepository = new FileUserRepository();
     public MessageRepository messageRepository = new FileMessageRepository();
 
     public UserService userService(){
-        return new JCFUserService(this.userRepository, this.channelService());
+        return new BasicUserService(this.userRepository, this.channelService());
     }
 
     public ChannelService channelService(){
-        return new JCFChannelService(this.channelRepository);
+        return new BasicChannelService(this.channelRepository);
     }
 
     public MessageService messageService(){
-        return new JCFMessageService(this.messageRepository);
+        return new BasicMessageService(this.messageRepository);
     }
-
-
-
 }
