+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service.jcf;
+package com.sprint.mission.discodeit.service.basic;
 
 import com.sprint.mission.discodeit.dto.AddUserDto;
 import com.sprint.mission.discodeit.entity.User;
@@ -10,13 +10,12 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-@Deprecated
-public class JCFUserService implements UserService {
+public class BasicUserService implements UserService {
 
     private final UserRepository userRepository;
     private final ChannelService channelService;
 
-    public JCFUserService(UserRepository userRepository, ChannelService channelService) {
+    public BasicUserService(UserRepository userRepository, ChannelService channelService) {
         this.userRepository = userRepository;
         this.channelService = channelService;
     }
